demos/backends/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-google-native>=0.7.0
PyYAML==6.0.1
